Split evalResultCache.ts into runSets.ts and trialEvents.ts (F-1411)
evalResultCache.ts was one line under the file-size gate, with no room
left for the unreadable-verdict tracking that follows. loadTrialEvents
(loads a trial's raw HTTP trace) and the run-set grouping helpers
(RunSet, groupRunSets, latestFailedRunSet, latestIncompleteRunSet) share
no shape or helper with the verdict.json read/write/scan logic that
stays behind, so they move to their own modules: hosted/trialEvents.ts
and hosted/runSets.ts. No behavior change; callers and tests now import
from the new paths.
